fixes (uart + sprite colision)

+++ proj/src/sprite/sprite.c
@@ -70,6 +70,7 @@ int sprite_colides(sprite_t *sprite1, sprite_t *sprite2)
 	int x2_s2 = sprite2->x + sprite2->img.width;
 	int y2_s2 = sprite2->y + sprite2->img.height;
 
-	return  x2_s1 > sprite2->x && sprite1->x < x2_s2 && y2_s1 > sprite2->y && sprite1->y < y2_s2;
+	return  x2_s1 > (int)sprite2->x && (int)sprite1->x < x2_s2 
+		&& y2_s1 > (int)sprite2->y && (int)sprite1->y < y2_s2;
 	// s1_x2 > s2_x && s1_x < s2_x2 && s1_y2 > s2_y && s1_y < s2_y2
 }
